Moved adding opus number functionality into its own module
